criando o adapter

+++ github/app/src/main/java/com/sinngjpeg/github/model/Repository.kt
@@ -4,12 +4,10 @@ data class Repository(
         val nomeRepository: String,
         val descricaoRepository: String,
         val imgFork: String?,
-        val numeroDeForks: Double,
+        val numeroDeForks: Int,
         val imgStar: String?,
-        val numeroDeStars: Double,
+        val numeroDeStars: Int,
         val imgPorfile: String?,
         val userName: String,
         val fullName: String
-
-
 )
